email_alerts/requirements.txt
psycopg2-binary
boto3
pytest
python-dotenv